requirements.txt
appdirs==1.4.4
catboost==1.2.7
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
graphviz==0.20.3
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.5.0
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
numba==0.60.0
numpy==1.26.4
nvidia-nccl-cu12==2.24.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
Pygments==2.19.1
pyod==2.0.3
pyparsing==3.2.1
pyshark==0.6
python-dateutil==2.9.0.post0
python-nmap==0.7.1
pytz==2024.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.15.0
six==1.17.0
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0
xgboost==2.1.3
